Add Deno task to run tests
Note that at the moment our tests run against a server running on port
8080, which will clash with the default port used when we run:

    deno task dev

We run the tests after waiting for a short delay, then after the tests
complete we run the shutdown task. Note the single "&" character towards
teh end of the test task. That tells Deno's shell that we want to run
the command to the left of it asynchronously, and immediately run `deno
task dev` without waiting for it to finish.

So what really happens is:

- We launch a command that sleeps for 5 seconds
- We launch a second command that runs `deno task dev`
- The first command then runs `deno test --allow-all` before calling the
  shutdown task

The shutdown task uses the `lsof` command to find the process ID of the
process that's listening on port 8080. That'll be the server we launched
with `deno task dev`.