Simplify and fix return code of 'deno task test'
Despite having tested the complicated one-line command that I wrote for
the test task, it didn't work properly. When the tests fail, it exits
with a successful exit code (i.e. 0).

I managed to edit the command so that it worked when run in a terminal
(which uses Bash), but when run via `deno task test` my new command
still didn't work. I think Deno has its own built-in shell command
parsing library, which seems to have different rules for working out the
exit code.

Moving all the steps to a custom script shell script makes it much
clearer what's going on, and means that we can reliably exit with the
correct exit code.

Why didn't I think of doing it like this in the first place?!