Tidy up HTML (and remove bootstrap)
I don't see the point of having a head.ejs partial outside of the
layout, so have inlined it.

I've removed Bootstrap because we won't be using it in this project.
In addition to removing the CSS file, I've removed the Bootstrap classes
from the HTML.

I've indented all the HTML by hand, as it appears as though VS Code
isn't doing that for us. We should look into that...